backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
simpy==4.1.1
numpy==1.26.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
aioredis==2.0.1
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6